// src/app/model/auth/authData.ts
export class AuthData {
    AccessToken: string = '';
    RefreshToken: string = '';
    DataRoute: string = '';
    ExpireAt: string = '';
    UserId: string = '';
    UserName: string = '';
    Role: number = 0;
    RoleName: string = '';
    DomainName: string = '';
    DomainKey: string = '';
    SubDomain: string = '';
    DomainKeyId: number = 0;
    Id: number = 0; 
}